+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Extent button now zooms to site markers
 -   Map zooms to sites when filters are submitted
 -   State filters are searchable
+-   Removed extra navbar buttons, changed style of Map button
 
 ### Fixed